% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-re-mcmc.R
\name{predict_mle_mcmc}
\alias{predict_mle_mcmc}
\title{Predict from an sdmTMB model with random effects sampled via MCMC}
\usage{
predict_mle_mcmc(
  object,
  model = c(1, 2),
  mcmc_iter = 500,
  mcmc_warmup = 250,
  print_stan_model = FALSE,
  stan_args = NULL
)
}
\arguments{
\item{object}{An \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}} model}

\item{model}{Which delta/hurdle model component?}

\item{mcmc_iter}{Iterations for MCMC residuals. Will take the last one.}

\item{mcmc_warmup}{Warmup for MCMC residuals.}

\item{print_stan_model}{Print the Stan model from MCMC residuals?}

\item{stan_args}{A list of arguments that will be passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}.}
}
\description{
Calculates predictions in link space on the original fitted data with random
effects estimated with MCMC via rstan/tmbstan while fixing fixed effects at
their MLE values. This avoids the Laplace approximation, which is useful for
checking residuals among other uses.
}
\examples{

library(sdmTMB)
mesh <- make_mesh(pcod_2011, c("X", "Y"), cutoff = 10)
fit <- sdmTMB(
  present ~ as.factor(year) + poly(depth, 3),
  data = pcod_2011, mesh = mesh,
  family = binomial()
)

# quick but can have issues because of Laplace approximation:
r1 <- residuals(fit, type = "mle-laplace")
qqnorm(r1)
qqline(r1)

# MCMC-based with fixed effects at MLEs; best but slowest:
set.seed(2938)
r2 <- residuals(fit, type = "mle-mcmc", mcmc_iter = 101, mcmc_warmup = 100)
qqnorm(r2)
qqline(r2)

# Example of passing control arguments to rstan::sampling():
# 11 iterations used for a quick example; don't do this normally
stan_args <- list(control = list(adapt_delta = 0.9, max_treedepth = 12))
r3 <- residuals(
  fit, type = "mle-mcmc", mcmc_iter = 11, mcmc_warmup = 10,
  stan_args = stan_args
)
}
